FU-155 refactor: separate server side http instance interceptors
